+++ infra/art-peace-infra/templates/consumer/consumer-configmap.yaml
@@ -10,3 +10,4 @@ data:
   CANVAS_FACTORY_CONTRACT_ADDRESS: {{ .Values.contracts.canvasFactory }}
   POSTGRES_PASSWORD: {{ .Values.postgres.password }}
   ROUND_NUMBER: {{ .Values.contracts.roundNumber }}
+  AVAIL_TURBO_API_KEY: {{ .Values.turboda.apiKey }}
